:memo: Move DatabaseList component folder
